refactor: first version of not needed torch

refactor: first version of not needed torch
