+++ docarray/document/mixins/proto.py
@@ -6,15 +6,6 @@ from docarray.document.base_node import BaseNode
 if TYPE_CHECKING:
     from docarray.proto import DocumentProto, NodeProto
 
-from docarray.typing import (
-    ID,
-    AnyUrl,
-    Embedding,
-    ImageUrl,
-    NdArray,
-    TextUrl,
-    TorchTensor,
-)
 
 T = TypeVar('T', bound='ProtoMixin')
 
@@ -23,6 +14,15 @@ class ProtoMixin(AbstractDocument, BaseNode):
     @classmethod
     def from_protobuf(cls: Type[T], pb_msg: 'DocumentProto') -> T:
         """create a Document from a protobuf message"""
+        from docarray.typing import (  # TorchTensor,
+            ID,
+            AnyUrl,
+            Embedding,
+            ImageUrl,
+            NdArray,
+            TextUrl,
+        )
+
         fields: Dict[str, Any] = {}
 
         for field in pb_msg.data:
@@ -34,6 +34,7 @@ class ProtoMixin(AbstractDocument, BaseNode):
             # the check should be delegated to the type level
             content_type_dict = dict(
                 ndarray=NdArray,
+                # torch_tensor=TorchTensor,
                 embedding=Embedding,
                 any_url=AnyUrl,
                 text_url=TextUrl,
