+++ docarray/typing/__init__.py
@@ -1,9 +1,12 @@
 from docarray.typing.id import ID
 from docarray.typing.tensor import NdArray, Tensor
+
+# from docarray.typing.tensor import TorchEmbedding, TorchTensor
 from docarray.typing.tensor.embedding import Embedding
 from docarray.typing.url import AnyUrl, ImageUrl, TextUrl
 
 __all__ = [
+    # 'TorchTensor',
     'NdArray',
     'Embedding',
     'ImageUrl',
@@ -11,6 +14,7 @@ __all__ = [
     'AnyUrl',
     'ID',
     'Tensor',
+    # 'TorchEmbedding',
 ]
 
 try:
