+++ docarray/typing/tensor/__init__.py
@@ -1,11 +1,18 @@
-from docarray.typing.tensor.embedding import Embedding, NdArrayEmbedding
+from docarray.typing.tensor.embedding import (  # , TorchEmbedding
+    Embedding,
+    NdArrayEmbedding,
+)
 from docarray.typing.tensor.ndarray import NdArray
 from docarray.typing.tensor.tensor import Tensor
 
+# from docarray.typing.tensor.torch_tensor import TorchTensor
+
 __all__ = [
     'NdArray',
+    # 'TorchTensor',
     'Tensor',
     'Embedding',
+    # 'TorchEmbedding',
     'NdArrayEmbedding',
 ]
