+++ docarray/typing/tensor/embedding/embedding.py
@@ -1,15 +1,7 @@
 from typing import Union
 
-from docarray.typing.tensor.embedding.ndarray import NdArrayEmbedding
+from docarray.typing.tensor.embedding.embedding_ndarray import NdArrayEmbedding
 
-try:
-    import torch  # noqa: F401
-except ImportError:
-    Embedding = Union[NdArrayEmbedding]  # type: ignore
+# from docarray.typing.tensor.embedding.embedding_torch import TorchEmbedding
 
-else:
-    from docarray.typing.tensor.embedding.torch import TorchEmbedding  # noqa: F401
-
-    Embedding = Union[NdArrayEmbedding, TorchEmbedding]  # type: ignore
-
-__all__ = ['Embedding']
+Embedding = Union[NdArrayEmbedding]  # , TorchEmbedding]
